requirements.txt
python-dotenv==0.13.0
uvicorn==0.18.3
fastapi==0.82.0
requests==2.25.1
bs4==0.0.1
pytz==2022.1
lxml==4.8.0
elasticsearch==7.10.1